Created shared-types file; cleanup
