+++ web/app/components/x/dropdown-list/toggle-action.ts
@@ -1,18 +1,12 @@
 import Component from "@glimmer/component";
+import { XDropdownListToggleComponentArgs } from "./_shared";
 
 interface XDropdownListToggleActionComponentSignature {
   Element: HTMLButtonElement;
-  Args: {
-    registerAnchor: (element: HTMLElement) => void;
-    onTriggerKeydown: (event: KeyboardEvent) => void;
-    toggleContent: () => void;
-    contentIsShown: boolean;
-    disabled?: boolean;
-    ariaControls: string;
-  };
+  Args: XDropdownListToggleComponentArgs;
   Blocks: {
     default: [];
-  }
+  };
 }
 
 export default class XDropdownListToggleActionComponent extends Component<XDropdownListToggleActionComponentSignature> {}
